refactor(prompt): move to subdirectory
